+++ heimdallm/bifrosts/sql/tests/sql/select/test_select.py
@@ -11,7 +11,7 @@ from .utils import PermissiveConstraints
 
 
 @dialects()
-def test_aliased_select_column(Bifrost: Type[Bifrost]):
+def test_aliased_select_column(dialect: str, Bifrost: Type[Bifrost]):
     """tests that we can select a column from a table that was aliased via a
     join"""
     query = """
@@ -32,12 +32,12 @@ def test_unqualified_columns():
 
     bifrost = Bifrost.mocked(PermissiveConstraints())
     with pytest.raises(exc.UnqualifiedColumn) as e:
-        bifrost.traverse(query)
+        bifrost.traverse(query, autofix=False)
     assert e.value.column == "col"
 
 
 @dialects()
-def test_unqualified_where_column(Bifrost: Type[Bifrost]):
+def test_unqualified_where_columns(dialect: str, Bifrost: Type[Bifrost]):
     """a column alias in the where class is valid, as long as the alias points to an
     actual aliased column. if it doesn't, we consider it unqualified."""
     query = """
@@ -50,7 +50,7 @@ LIMIT 5;
 
     bifrost = Bifrost.mocked(PermissiveConstraints())
     with pytest.raises(exc.UnqualifiedColumn) as e:
-        bifrost.traverse(query)
+        bifrost.traverse(query, autofix=False)
     assert e.value.column == "customer_id"
 
 
@@ -64,7 +64,7 @@ LIMIT 5;
         'select t1.col from t1 join "t2" on t1.jid = "t2".jid',
     ),
 )
-def test_escapes(Bifrost: Type[Bifrost], query):
+def test_escapes(dialect: str, Bifrost: Type[Bifrost], query):
     class MyConstraints(PermissiveConstraints):
         def select_column_allowed(self, column: FqColumn):
             return column.name in {"t1.col"}
@@ -77,7 +77,7 @@ def test_escapes(Bifrost: Type[Bifrost], query):
 
 
 @dialects()
-def test_disallowed_select_column(Bifrost: Type[Bifrost]):
+def test_disallowed_select_column(dialect: str, Bifrost: Type[Bifrost]):
     """tests that our constraint validator works when trying to select a
     column"""
     query = "select t1.col from t1"
@@ -102,7 +102,7 @@ def test_disallowed_select_column(Bifrost: Type[Bifrost]):
 
 
 @dialects()
-def test_required_constraint(Bifrost: Type[Bifrost]):
+def test_required_constraint(dialect: str, Bifrost: Type[Bifrost]):
     """tests that our required constraint restricts the query correctly"""
 
     # missing constraint
@@ -137,7 +137,7 @@ def test_required_constraint(Bifrost: Type[Bifrost]):
     # impossible for us to trace back to the original table
     query = "select t1.col from t1 as aliased where id=:id"
     with pytest.raises(exc.UnqualifiedColumn):
-        bifrost.traverse(query)
+        bifrost.traverse(query, autofix=False)
 
 
 @dialects()
@@ -148,7 +148,7 @@ def test_required_constraint(Bifrost: Type[Bifrost]):
         "select `` fro t1",  # just broken
     ),
 )
-def test_broken_query(Bifrost: Type[Bifrost], query):
+def test_broken_query(dialect: str, Bifrost: Type[Bifrost], query):
     """tests that we raise an exception when we cannot parse the query"""
 
     bifrost = Bifrost.mocked(PermissiveConstraints())
@@ -157,7 +157,7 @@ def test_broken_query(Bifrost: Type[Bifrost], query):
 
 
 @dialects()
-def test_select_column_arith(Bifrost: Type[Bifrost]):
+def test_select_column_arith(dialect: str, Bifrost: Type[Bifrost]):
     query = "select t1.col + 1 as plus_one from t1"
     bifrost = Bifrost.mocked(PermissiveConstraints())
     bifrost.traverse(query)
@@ -171,14 +171,14 @@ def test_select_column_arith(Bifrost: Type[Bifrost]):
         "select t1.id, (1 + 1) as two from t1",
     ),
 )
-def test_select_expr(Bifrost: Type[Bifrost], query):
+def test_select_expr(dialect: str, Bifrost: Type[Bifrost], query):
     """select a non-column expression"""
     bifrost = Bifrost.mocked(PermissiveConstraints())
     bifrost.traverse(query)
 
 
 @dialects()
-def test_conflicting_validations(Bifrost: Type[Bifrost]):
+def test_conflicting_validation(dialect: str, Bifrost: Type[Bifrost]):
     query = "select t1.col from t1"
 
     class MyConstraints(PermissiveConstraints):
@@ -193,7 +193,7 @@ def test_conflicting_validations(Bifrost: Type[Bifrost]):
 
 
 @dialects()
-def test_count_disallowed_column(Bifrost: Type[Bifrost]):
+def test_count_disallowed_column(dialect: str, Bifrost: Type[Bifrost]):
     """we're allowed to count a disallowed column"""
 
     class GeneralConstraints(PermissiveConstraints):