+++ heimdallm/bifrosts/sql/tests/sql/select/test_validators.py
@@ -31,7 +31,7 @@ class Succeed2(PermissiveConstraints):
 
 
 @dialects()
-def test_all_fail(Bifrost: Type[Bifrost]):
+def test_all_fail(dialect: str, Bifrost: Type[Bifrost]):
     bifrost = Bifrost.mocked([Fail1(), Fail2()])
 
     query = """
@@ -45,7 +45,7 @@ def test_all_fail(Bifrost: Type[Bifrost]):
 
 
 @dialects()
-def test_all_fail_order(Bifrost: Type[Bifrost]):
+def test_all_fail_order(dialect: str, Bifrost: Type[Bifrost]):
     """we always raise the last error, switching the validators proves this."""
     bifrost = Bifrost.mocked([Fail2(), Fail1()])
 
@@ -60,7 +60,7 @@ def test_all_fail_order(Bifrost: Type[Bifrost]):
 
 
 @dialects()
-def test_one_succeed(Bifrost: Type[Bifrost]):
+def test_one_succeed(dialect: str, Bifrost: Type[Bifrost]):
     bifrost = Bifrost.mocked([Fail1(), Succeed1()])
 
     query = """
@@ -72,7 +72,7 @@ def test_one_succeed(Bifrost: Type[Bifrost]):
 
 
 @dialects()
-def test_both_succeed(Bifrost: Type[Bifrost]):
+def test_both_succeed(dialect: str, Bifrost: Type[Bifrost]):
     bifrost = Bifrost.mocked([Succeed1(), Succeed2()])
 
     query = """